# paas-mediation-service/docs/swagger.yaml
definitions:
...
  v2.AnnotationResource:
    properties:
      annotationValue:
        type: string
      namespace:
        type: string
      resourceName:
        type: string
    type: object
  v2.AppVersionData:
    properties:
      appName:
        type: string
      appVersion:
        type: string
      deployTime:
        type: string
    type: object
...
  v2.DeploymentFamilyVersion:
    properties:
      app_name:
        type: string
      app_version:
        type: string
      bluegreen_version:
        type: string
      family_name:
        type: string
      name:
        type: string
      state:
        type: string
      version:
        type: string
    type: object
  v2.DeploymentResponse:
    properties:
      deployments:
        items:
          additionalProperties:
            $ref: '#/definitions/v2.DeploymentRollout'
          type: object
        type: array
      pod_status_websocket:
        type: string
    type: object
  v2.DeploymentRollout:
    properties:
      active:
        type: string
      kind:
        type: string
      rolling:
        type: string
    type: object
...
  v2.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  v2.Metadata:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      generation:
        type: integer
      kind:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      namespace:
        type: string
      resourceVersion:
        type: string
      uid:
        type: string
    type: object
  v2.Namespace:
    properties:
      metadata:
        $ref: '#/definitions/v2.Metadata'
    type: object
...
  v2.RolloutDeploymentBody:
    properties:
      deployment_names:
        items:
          type: string
        type: array
      parallel:
        type: boolean
    type: object
...
info:
  description: API for Paas Mediation.
  title: Paas Mediation API
  version: "2.0"
paths:
...
  /api/v2/namespaces:
    get:
      consumes:
      - application/json
      description: Get namespaces
      operationId: v2-get-namespaces
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v2.Namespace'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get namespaces
      tags:
      - since:2.0
  /api/v2/namespaces/{namespace}/annotations:
    get:
      consumes:
      - application/json
      description: Get resources by resource type and annotation name in namespace
      operationId: v2-get-annotationresource
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource type
        in: query
        name: resourceType
        required: true
        type: string
      - description: annotation name
        in: query
        name: annotation
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v2.AnnotationResource'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get resources by resource type and annotation name
      tags:
      - since:2.0
...
  /api/v2/namespaces/{namespace}/configmaps/versions:
    get:
      consumes:
      - application/json
      description: Get versions from 'version' ConfigMap
      operationId: v2-get-versions
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v2.AppVersionData'
            type: array
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get versions from 'version' ConfigMap
      tags:
      - since:2.0
  /api/v2/namespaces/{namespace}/deployment-family/{family_name}:
    get:
      consumes:
      - application/json
      description: Get DeploymentFamily data based on Deployments labeled with 'family_name'
        label with value specified via 'deployment-family' path param
      operationId: v2-get-deploymentfamily-versions
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: family name
        in: path
        name: family_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v2.DeploymentFamilyVersion'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get DeploymentFamily data based on Deployments labeled with 'family_name'
        label with value specified via 'deployment-family' path param
      tags:
      - since:2.0
...
  /api/v2/namespaces/{namespace}/rollout:
    post:
      consumes:
      - application/json
      description: Restart Deployments in bulk by names in namespace in parallel or
        sequentially
      operationId: v2-post-restartdeployments-bulk
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v2.RolloutDeploymentBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.DeploymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Restart Deployments in bulk by names in namespace in parallel or sequentially
      tags:
      - since:2.0
  /api/v2/namespaces/{namespace}/rollout/{resource-name}:
    post:
      consumes:
      - application/json
      description: RestartDeployment by name in namespace
      operationId: v2-post-restartdeployment
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource name
        in: path
        name: resource-name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.DeploymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: RestartDeployment
      tags:
      - since:2.0
...
produces:
- application/json
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...